third_party/blink/web_tests/fast/dom/shadow/cppevent-in-shadow-expected.txt

dark: copy is fired.
host: copy is fired.
dark: cut is fired.
host: cut is fired.
dark: paste is fired.
host: paste is fired.
PASS successfullyParsed is true

TEST COMPLETE
